# contracts/http/openapi.yaml
openapi: 3.1.0
info:
  title: VPN Subscription Platform API
  version: 0.7.1-stage7
  license:
    name: UNLICENSED
    identifier: LicenseRef-UNLICENSED
  description: >
    Stage 7 contract. Durable Telegram notification delivery and the bounded
    mTLS administrator API/CLI are executable locally. Production deployment
    and credential issuance remain out of scope.
servers:
  - url: https://api.example.invalid
    description: Public API placeholder
  - url: https://internal.example.invalid
    description: Internal API placeholder requiring mTLS service identity
  - url: https://admin.example.invalid
    description: Admin API placeholder requiring admin mTLS identity
  - url: https://sub.example.invalid
    description: Happ subscription endpoint placeholder
security:
  - ServiceMTLS: []
...
paths:
...
  /internal/v1/dead-letters:
    get:
      tags: [Admin]
      summary: List sanitized notification DLQ metadata
      operationId: listNotificationDeadLetters
      x-contract-status: implemented-stage7
      x-allowed-service-identities: [admin-service]
      parameters:
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: Coordinates, digest, and bounded reason only
          content:
            application/json:
              schema: {$ref: '#/components/schemas/NotificationDeadLetterList'}
        '400': {$ref: '#/components/responses/Error'}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    ServiceMTLS:
      type: mutualTLS
      description: >
        Internal service identity from a trusted platform CA. Authorization is
        based on a strict SPIFFE URI from the verified certificate chain and
        per-endpoint allowed identities.
    ProvisioningMTLS:
      type: mutualTLS
      description: >
        Internal mTLS identity restricted to provisioning-service for fetching
        sensitive credential material from access-service.
    NodeAgentMTLS:
      type: mutualTLS
      description: >
        Private management-plane mTLS. The client must be provisioning-service,
        and provisioning-service additionally verifies the exact registered
        node SPIFFE URI in the server's verified certificate chain.
    AdminMTLS:
      type: mutualTLS
      description: >
        Admin CLI/client mTLS identity mapped to RBAC roles and audited actions.
    TelegramWebhookSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Telegram webhook secret token configured through Bot API setWebhook.
    SubscriptionUrlBearer:
      type: http
      scheme: bearer
      bearerFormat: URL path token
      description: >
        Bearer-secret semantics for `GET /s/{token}`. Happ v1 carries this
        token in the path, not in the Authorization header.
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
  responses:
    Error:
      description: Safe error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [code, message, request_id]
      additionalProperties: false
      properties:
        code:
          type: string
          maxLength: 128
        message:
          type: string
          maxLength: 512
        request_id:
          type: string
          maxLength: 128
        fields:
          type: array
          items:
            type: object
            required: [field, code]
            additionalProperties: false
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
...
    NotificationDeadLetterList:
      type: object
      required: [dead_letters]
      additionalProperties: false
      properties:
        dead_letters:
          type: array
          items: {$ref: '#/components/schemas/NotificationDeadLetter'}
    NotificationDeadLetter:
      type: object
      required: [source_topic, source_partition, source_offset, payload_sha256, reason_code, state, created_at, updated_at]
      additionalProperties: false
      properties:
        source_topic: {type: string, maxLength: 249}
        source_partition: {type: integer, minimum: 0}
        source_offset: {type: integer, format: int64, minimum: 0}
        payload_sha256: {type: string, pattern: '^[0-9a-f]{64}$'}
        event_type: {type: string, maxLength: 128}
        reason_code: {type: string, maxLength: 64}
        state: {type: string, enum: [available, replay_requested, replayed]}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...